Registration And Login with JWT and ownership check

+++ WebApplication1/Controllers/V1/PostsController.cs
@@ -7,6 +7,7 @@ using Post_Surfer.Services;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Post_Surfer.Extensions;
 
 // For more information on enabling MVC for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -26,7 +27,11 @@ namespace Post_Surfer.Controllers.V1
         [HttpPost(APIRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name,Id= Guid.NewGuid() };
+            var post = new Post
+            {
+                Name = postRequest.Name, Id = Guid.NewGuid(),
+                UserId = HttpContext.GetUserId()
+            };
             await _postService.CreatePostAsync(post);
             var baseUrl = $"{HttpContext.Request.Scheme}://{HttpContext.Request.Host.ToUriComponent()}";
             //var locationUrl = baseUrl + "/" + APIRoutes.Posts.Create.Replace("postId", post.Id);
@@ -41,7 +46,11 @@ namespace Post_Surfer.Controllers.V1
             {
                 return BadRequest();
             }
-
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own the post" });
+            }
             var status = await _postService.DeletePostAsync(postId);
 
             if (status)
@@ -89,7 +98,13 @@ namespace Post_Surfer.Controllers.V1
                 return BadRequest();
             }
 
-            var post = new Post { Id = postRequest.Id, Name = postRequest.Name };
+            var userOwnsPost = await _postService.UserOwnsPostAsync(postRequest.Id, HttpContext.GetUserId());
+            if (!userOwnsPost)
+            {
+                return BadRequest(new { error = "You do not own the post" });
+            }
+            var post = await _postService.GetPostByIdAsync(postRequest.Id);
+            post.Name = postRequest.Name;
             var status = await _postService.UpdatePostAsync(post);
             if (status)
             {
